refactor: separate all the context into a individual file
