+++ packages/extension-polkagate/tsconfig.build.json
@@ -18,8 +18,6 @@
     { "path": "../extension-inject/tsconfig.build.json"
     },
     { "path": "../extension-mocks/tsconfig.build.json"
-    },
-    { "path": "../extension-ui/tsconfig.build.json"
     }
   ]
 }
