+++ packages/extension-ui/src/Popup/contexts/index.ts
@@ -1,0 +1,18 @@
+// Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { default as AccountAssetProvider } from './AccountAssetProvider';
+export { default as AccountIconThemeProvider } from './AccountIconThemeProvider';
+export { default as AccountProvider } from './AccountProvider';
+export { default as ActionProvider } from './ActionProvider';
+export { default as AlertProvider } from './AlertProvider';
+export { default as ApiProvider } from './ApiProvider';
+export { default as CurrencyProvider } from './CurrencyProvider';
+export { default as FetchingProvider } from './FetchingProvider';
+export { default as GenesisHashOptionsProvider } from './GenesisHashOptionsProvider';
+export { default as MediaProvider } from './MediaProvider';
+export { default as ReferendaProvider } from './ReferendaProvider';
+export { default as RequestsProvider } from './RequestsProvider';
+export { default as SettingsProvider } from './SettingsProvider';
+export { default as UserAddedChainsProvider } from './UserAddedChainsProvider';
+export { default as WorkerProvider } from './WorkerProvider';
